Revert d132063..fdfe887
This rolls back to commit d1320634f707b1adc9d031794be0d7c116776485.

+++ delete.php
@@ -1,16 +1,19 @@
 <?php
 	include('config.php');
+
+	if (isset($_POST['delete'])) {
 		
-	// get value from hidden field in form
-	$cid = $_POST['clientRecordId'];
-	
-	$sql_delete = "DELETE FROM test WHERE id='$cid'";
-	
-	// tells if added successfully or not
-	if (!mysqli_query($con,$sql_delete)) {
-		die('Error: ' . mysqli_error($con));
+		// get value from hidden field in form
+		$cid = $_POST['clientRecordId'];
+		
+		$sql_delete = "DELETE FROM test WHERE id='$cid'";
+		
+		// tells if added successfully or not
+		if (!mysqli_query($con,$sql_delete)) {
+			die('Error: ' . mysqli_error($con));
+		}
+		echo "1 record deleted";
 	}
-	echo "Entry deleted";
 	
 	// close connection 
 	mysqli_close($con);
